move settings & remove "common" from core
