+++ app/src/main/java/com/realityexpander/tasky/auth_feature/data/repository/local/entities/AuthInfoEntity.kt
@@ -1,8 +1,8 @@
 package com.realityexpander.tasky.auth_feature.data.repository.local.entities
 
-import com.realityexpander.tasky.core.common.AuthToken
-import com.realityexpander.tasky.core.common.UserId
-import com.realityexpander.tasky.core.common.Username
+import com.realityexpander.tasky.core.util.AuthToken
+import com.realityexpander.tasky.core.util.UserId
+import com.realityexpander.tasky.core.util.Username
 
 data class AuthInfoEntity(
     val authToken: AuthToken? = null,
