+++ app/src/main/java/com/realityexpander/tasky/auth_feature/data/repository/remote/authApiImpls/AuthApiFakeImpl.kt
@@ -2,7 +2,7 @@ package com.realityexpander.tasky.auth_feature.data.repository.remote.authApiImp
 
 import com.realityexpander.tasky.auth_feature.data.repository.remote.DTOs.auth.AuthInfoDTO
 import com.realityexpander.tasky.auth_feature.data.repository.remote.IAuthApi
-import com.realityexpander.tasky.core.common.*
+import com.realityexpander.tasky.core.util.*
 import kotlinx.coroutines.delay
 import javax.inject.Inject
