+++ app/src/main/java/com/realityexpander/tasky/auth_feature/data/repository/remote/authApiImpls/AuthApiImpl.kt
@@ -3,7 +3,7 @@ package com.realityexpander.tasky.auth_feature.data.repository.remote.authApiImp
 import com.realityexpander.tasky.auth_feature.data.repository.remote.DTOs.auth.ApiCredentialsDTO
 import com.realityexpander.tasky.auth_feature.data.repository.remote.DTOs.auth.AuthInfoDTO
 import com.realityexpander.tasky.auth_feature.data.repository.remote.IAuthApi
-import com.realityexpander.tasky.core.common.*
+import com.realityexpander.tasky.core.util.*
 import com.realityexpander.tasky.core.data.remote.utils.getErrorBodyMessage
 import retrofit2.HttpException
 import javax.inject.Inject
