+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/AuthInfo.kt
@@ -1,9 +1,9 @@
 package com.realityexpander.tasky.auth_feature.domain
 
 import android.os.Parcelable
-import com.realityexpander.tasky.core.common.AuthToken
-import com.realityexpander.tasky.core.common.UserId
-import com.realityexpander.tasky.core.common.Username
+import com.realityexpander.tasky.core.util.AuthToken
+import com.realityexpander.tasky.core.util.UserId
+import com.realityexpander.tasky.core.util.Username
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
