+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/IAuthRepository.kt
@@ -3,10 +3,10 @@ package com.realityexpander.tasky.auth_feature.domain
 import com.realityexpander.tasky.auth_feature.domain.validation.ValidateEmail
 import com.realityexpander.tasky.auth_feature.domain.validation.ValidatePassword
 import com.realityexpander.tasky.auth_feature.domain.validation.ValidateUsername
-import com.realityexpander.tasky.core.common.AuthToken
-import com.realityexpander.tasky.core.common.Email
-import com.realityexpander.tasky.core.common.Password
-import com.realityexpander.tasky.core.common.Username
+import com.realityexpander.tasky.core.util.AuthToken
+import com.realityexpander.tasky.core.util.Email
+import com.realityexpander.tasky.core.util.Password
+import com.realityexpander.tasky.core.util.Username
 
 
 interface IAuthRepository {
