+++ app/src/main/java/com/realityexpander/tasky/core/util/Exceptions.kt
@@ -1,4 +1,4 @@
-package com.realityexpander.tasky.core.common
+package com.realityexpander.tasky.core.util
 
 object Exceptions {
